Make activity log references to additional _id fields bignint Remove admin_id from the attributes shown in the parent item of activity log blocks

+++ db/table_generators/activity_logs_table.rb
@@ -69,7 +69,7 @@ EOF
       attrib_pair = {}
       attrib.each do |a|
         f = 'varchar'
-        f = 'integer' if a.end_with?('_id')
+        f = 'bigint' if a.end_with?('_id')
         f = 'date' if a.end_with?('_when')
         f = 'date' if a.end_with?('_date')
         f = 'varchar' if a == 'data'
